Improve efficiency
Piping to AWK to parse the CSV file is inefficient. For every pipe to AWK we're forking a new shell, executing awk and then returning. We can handle this processing directly within the script.

Assuming the CSV file is formatted something like:
1password,1student_number,1username
2password,2student_number,2username
So on so forth.

The delimiter to be , in order to allow us to parse the CSV file.